Administrator can see user's expenses in user view(index and edit)

+++ database/migrations/2017_08_21_181446_create_expenses_table.php
@@ -17,7 +17,7 @@ class CreateExpensesTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned();
             $table->string('name');
-            $table->float('amount', 8, 2);
+            $table->integer('amount');
             $table->timestamps();
         });
